+++ src/nns/node/run/cache.rs
@@ -1,6 +1,8 @@
-use crate::nns::{leaf::NnsLeafCacheRequest, node::report::NnsNodeCacheRequest};
-use std::path::Path;
+use crate::nns::{
+    NnsCommandError, command_icp_root, leaf::NnsLeafCacheRequest, node::report::NnsNodeCacheRequest,
+};
 
-pub(super) fn cache_request(icp_root: &Path, network: &str) -> NnsNodeCacheRequest {
-    NnsNodeCacheRequest::from_root_network(icp_root, network)
+pub(super) fn cache_request(network: &str) -> Result<NnsNodeCacheRequest, NnsCommandError> {
+    let icp_root = command_icp_root()?;
+    Ok(NnsNodeCacheRequest::from_root_network(&icp_root, network))
 }